+++ oauth2/src/test/scala/org/broadinstitute/dsde/workbench/oauth2/mock/FakeOpenIDConnectConfiguration.scala
@@ -1,13 +1,25 @@
 package org.broadinstitute.dsde.workbench.oauth2.mock
 
-import org.broadinstitute.dsde.workbench.oauth2.{ClientId, OpenIDConnectConfiguration, OpenIDProviderMetadata}
+import org.broadinstitute.dsde.workbench.oauth2.{
+  ClientId,
+  OpenIDConnectConfiguration,
+  OpenIDProviderMetadata,
+  OpenIdProvider
+}
+import org.http4s.Uri
 
 class FakeOpenIDConnectConfiguration extends OpenIDConnectConfiguration {
   override def clientId: ClientId = ClientId("some-client")
 
-  override def authorityEndpoint: String = "https://fake"
+  override def openIdProvider: OpenIdProvider = OpenIdProvider(
+    authorityEndpoint,
+    Uri.fromString("https://fake/.well-known/openid-configuration").toOption.get,
+    providerMetadata
+  )
 
-  override def providerMetadata: OpenIDProviderMetadata =
+  val authorityEndpoint: String = "https://fake"
+
+  val providerMetadata: OpenIDProviderMetadata =
     OpenIDProviderMetadata("fake-issuer", "fake-authorize", "fake-token", Option("fake-end-session"))
 
   override def processAuthorizeQueryParams(params: Seq[(String, String)]): Seq[(String, String)] = params