+++ oauth2/CHANGELOG.md
@@ -3,9 +3,12 @@
 This file documents changes to the `workbench-oauth2` library, including notes on how to upgrade to new versions.
 
 ## 0.7
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.7-d2b30c4"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.7-TRAVIS-REPLACE-ME"`
 Changed:
 - removed support for google oauth2, only azure b2c authentication is supported now in Terra UIs and swagger-ui
+
+Added:
+- token replacement in openapi.yaml for OPEN_ID_CONNECT_URL, OAUTH_AUTHORIZATION_URL, and OAUTH_TOKEN_URL all with optional _WITH_GOOGLE_BILLING_SCOPE suffix
 
 ## 0.6
 SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.6-d314413"`